+++ src/structures/operations/indexscan.h
@@ -14,9 +14,9 @@ class IndexScan : public Operation{
             }
         }
 
-        virtual float getStartCost(Database* database);
-        virtual float getRuntimeCost(Database* database);
-        virtual float getNt();
+        virtual double getStartCost(Database* database);
+        virtual double getRuntimeCost(Database* database);
+        virtual double getNt();
 
     private:
         Index* index;
