+++ src/structures/operations/nested-join.h
@@ -7,9 +7,9 @@
 class NestedJoin : public Operation{
     public:
         NestedJoin(bool foreign_key = false);
-        virtual float getStartCost(Database* database);
-        virtual float getRuntimeCost(Database* database);
-        virtual float getNt();
+        virtual double getStartCost(Database* database);
+        virtual double getRuntimeCost(Database* database);
+        virtual double getNt();
     private:
         int nt;
 };
